requirements.txt
opencv-python
speechrecognition
pyaudio
google-generativeai
keyboard
gTTS
playsound
Flask
SocketIO
re
